reviews_service/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
PyJWT
passlib[bcrypt]
memory_profiler